replace Email::Simple with regex
We only needed this for "key: value" lines at the top, before a blank.
Email::Simple was complaining because we assume the files on disk are
UTF-8, then decode them before passing them to Email::Simple, which
expects octets (and ASCII ones, at that!).

Instead, let's just parse the "email" on our own.  We don't have any
encoding other than:

1. "key: value" pairs until the first run of blank lines
2. then the body